extract test details into page
